requirements.txt
certifi==2019.6.16
Django==2.2.5
django-multi-form-view==2.0.1
gunicorn==19.9.0
olefile==0.46
Pillow==6.1.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
django-bootstrap4==1.1.1
djangorestframework==3.10.3
stripe==2.44.0
